% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_MetaRegulon_score.r
\name{cal_MetaRegulon}
\alias{cal_MetaRegulon}
\title{Calculate the MetaRegulon score for each MetaModule in cell subtypes/metabolic subgtypes/sample conditions.}
\usage{
cal_MetaRegulon(
  object,
  feature,
  state,
  interested_MM,
  MM_list,
  markers,
  lisa_file,
  ligand_target_matrix,
  lr_network,
  sample_tech,
  output_path,
  RP_score,
  file_name
)
}
\arguments{
\item{object}{A Seurat object}

\item{feature}{The feature used to cluster cells, typically chosen from the column names of the meta.data in the Seurat object}

\item{state}{The cell subtype, metabolic subtype, or sample condition}

\item{interested_MM}{The name of the metabolic reaction or interested gene set}

\item{MM_list}{A list of metabolic reactions or interested gene sets}

\item{markers}{Marker genes enriched in the corresponding state}

\item{lisa_file}{The path to the LISA results for the marker genes}

\item{ligand_target_matrix}{The path to the ligand-target matrix (provided)}

\item{lr_network}{The ligand-receptor network matrix (provided)}

\item{sample_tech}{The technology used for sequencing}

\item{output_path}{The path to save the MetaRegulon results}

\item{RP_score}{The path to the RP results (provided)}

\item{file_name}{The file name for the MetaRegulon results}
}
\description{
Calculate the MetaRegulon score for each MetaModule in cell subtypes/metabolic subgtypes/sample conditions.
}
