% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_MetaRegulon_score.r
\name{cal_target_score}
\alias{cal_target_score}
\title{cal_target_score}
\usage{
cal_target_score(lisa_res_in_metroscreen, factor, data, RP_score)
}
\arguments{
\item{lisa_res_in_metroscreen}{A seurat object}

\item{factor}{Feature used to cluster cells, usually use the colnames of the meta.data of the seurat object}

\item{data}{State of the interested cell types pr metabolic state}

\item{RP_score}{}
}
\description{
cal_target_score
}
