% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{doheatmap_feature}
\alias{doheatmap_feature}
\title{Draw the heatmap to show the difference.}
\usage{
doheatmap_feature(object, sample_info, feature, width, height, cols)
}
\arguments{
\item{object}{A matrix, the row is the MetaModule and the column is the samples/Metacells/cell, the value is the MetaModule score}

\item{sample_info}{A vector including ample information for the samples/Metacells/cells}

\item{feature}{Feature to show}

\item{width}{Width of the figure}

\item{height}{Height of the figure}
}
\description{
Draw the heatmap to show the difference.
}
