% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{FindAllMarkers_MetaModule}
\alias{FindAllMarkers_MetaModule}
\title{Find the differentially enriched MetaModule.}
\usage{
FindAllMarkers_MetaModule(object, sample_info, sample_tech)
}
\arguments{
\item{object}{A matrix, the row is the MetaModule and the column is the  samples/Metacells/cell, the value is the MetaModule score}

\item{sample_info}{A vector including ample information for the  samples/Metacells/cell}

\item{sample_tech}{scRNA or bulk.}
}
\description{
Find the differentially enriched MetaModule.
}
