% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_MetaRegulon_score.r
\name{gene_gene_interaction}
\alias{gene_gene_interaction}
\title{gene_gene_interaction}
\usage{
gene_gene_interaction(output_path, file_name, file_res, MM_list, interested_MM)
}
\arguments{
\item{output_path}{}

\item{file_name}{}

\item{file_res}{}

\item{interested_MM}{}
}
\description{
gene_gene_interaction
}
