% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_MetaRegulon_score.r
\name{get_causal}
\alias{get_causal}
\title{get_causal}
\usage{
get_causal(
  data,
  tr_activity,
  lr_activity,
  factor,
  MM_list,
  pathway_index,
  times,
  factor_class
)
}
\arguments{
\item{lr_activity}{}

\item{factor_class}{}
}
\description{
get_causal
}
