% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_MetaRegulon_score.r
\name{ligand_target_interaction}
\alias{ligand_target_interaction}
\title{ligand_target_interaction}
\usage{
ligand_target_interaction(
  output_path,
  file_name,
  state,
  file_res,
  markers,
  MM_list,
  interested_MM,
  ligand_target_matrix,
  lr_network
)
}
\arguments{
\item{output_path}{}

\item{file_name}{}

\item{state}{}

\item{lr_network}{}
}
\description{
ligand_target_interaction
}
