% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_MetaRegulon_score.r
\name{repref_each}
\alias{repref_each}
\title{repref_each}
\usage{
repref_each(
  tr_cor,
  tr_activity,
  lr_activity,
  ligand_cor,
  gene_cor,
  pathway_index,
  MM_list,
  data,
  output_path,
  file_name
)
}
\arguments{
\item{tr_activity}{}

\item{lr_activity}{}

\item{file_name}{}
}
\description{
repref_each
}
