% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_MetaRegulon_score.r
\name{tr_target_interaction}
\alias{tr_target_interaction}
\title{tr_target_interaction}
\usage{
tr_target_interaction(
  output_path,
  file_name,
  lisa_file,
  file_res,
  RP_score,
  MM_list,
  interested_MM
)
}
\arguments{
\item{lisa_file}{}

\item{interested_MM}{}
}
\description{
tr_target_interaction
}
